// src/catalogi-generated-client/1.3.1/api/data-contracts.ts
export interface CheckListItem {
  /**
   * itemnaam
   * De betekenisvolle benaming van het checklistitem
   * @maxLength 30
   */
  itemnaam: string;
  /**
   * toelichting
   * Beschrijving van de overwegingen bij het controleren van het aandachtspunt
   * @maxLength 1000
   */
  toelichting?: string | null;
  /**
   * vraagstelling
   * Een betekenisvolle vraag waaruit blijkt waarop het aandachtspunt gecontroleerd moet worden.
   * @maxLength 255
   */
  vraagstelling: string;
  /**
   * verplicht
   * Het al dan niet verplicht zijn van controle van het aandachtspunt voorafgaand aan het bereiken van de status van het gerelateerde STATUSTYPE.
   */
  verplicht?: boolean;
}
// ...
export interface FieldValidationError {
  /**
   * name
   * Naam van het veld met ongeldige gegevens
   */
  name: string;
  /**
   * code
   * Systeemcode die het type fout aangeeft
   */
  code: string;
  /**
   * reason
   * Uitleg wat er precies fout is met de gegevens
   */
  reason: string;
}

export enum FormaatEnum {
  Tekst = "tekst",
  Getal = "getal",
  Datum = "datum",
  DatumTijd = "datum_tijd",
}

/** Formaat van HTTP 4xx en 5xx fouten. */
export interface Fout {
  /**
   * type
   * URI referentie naar het type fout, bedoeld voor developers
   */
  type?: string;
  /**
   * code
   * Systeemcode die het type fout aangeeft
   */
  code: string;
  /**
   * title
   * Generieke titel voor het type fout
   */
  title: string;
  /**
   * status
   * De HTTP status code
   */
  status: number;
  /**
   * detail
   * Extra informatie bij de fout, indien beschikbaar
   */
  detail: string;
  /**
   * instance
   * URI met referentie naar dit specifiek voorkomen van de fout. Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.
   */
  instance: string;
}
// ...
export interface PaginatedStatusTypeList {
  /** @example 123 */
  count?: number;
  /**
   * @format uri
   * @example "http://api.example.org/accounts/?page=4"
   */
  next?: string | null;
  /**
   * @format uri
   * @example "http://api.example.org/accounts/?page=2"
   */
  previous?: string | null;
  results?: StatusType[];
}
// ...
export interface PatchedStatusType {
  /**
   * url
   * URL-referentie naar dit object. Dit is de unieke identificatie en locatie van dit object.
   * @format uri
   * @minLength 1
   * @maxLength 1000
   */
  url?: string;
  /**
   * omschrijving
   * Een korte, voor de initiator van de zaak relevante, omschrijving van de aard van de STATUS van zaken van een ZAAKTYPE.
   * @maxLength 80
   */
  omschrijving?: string;
  /**
   * omschrijving generiek
   * Algemeen gehanteerde omschrijving van de aard van STATUSsen van het STATUSTYPE
   * @maxLength 80
   */
  omschrijvingGeneriek?: string;
  /**
   * statustekst
   * De tekst die wordt gebruikt om de Initiator te informeren over het bereiken van een STATUS van dit STATUSTYPE bij het desbetreffende ZAAKTYPE.
   * @maxLength 1000
   */
  statustekst?: string;
  /**
   * Zaaktype
   * URL-referentie naar het ZAAKTYPE van ZAAKen waarin STATUSsen van dit STATUSTYPE bereikt kunnen worden.
   * @format uri
   */
  zaaktype?: string;
  /**
   * catalogus
   * @format uri
   */
  catalogus?: string;
  /**
   * Zaaktype
   * Unieke identificatie van het ZAAKTYPE binnen de CATALOGUS waarin het ZAAKTYPE voorkomt.
   */
  zaaktypeIdentificatie?: string;
  /**
   * statustypevolgnummer
   * Een volgnummer voor statussen van het STATUSTYPE binnen een zaak.
   * @min 1
   * @max 9999
   */
  volgnummer?: number;
  /**
   * isEindstatus
   * Geeft aan dat dit STATUSTYPE een eindstatus betreft. Dit gegeven is afgeleid uit alle STATUSTYPEn van dit ZAAKTYPE met het hoogste volgnummer.
   */
  isEindstatus?: boolean;
  /**
   * informeren
   * Aanduiding die aangeeft of na het zetten van een STATUS van dit STATUSTYPE de Initiator moet worden geïnformeerd over de statusovergang.
   */
  informeren?: boolean;
  /**
   * doorlooptijd status
   * De door de zaakbehandelende organisatie(s) gestelde norm voor de doorlooptijd voor het bereiken van STATUSsen van dit STATUSTYPE bij het desbetreffende ZAAKTYPE.
   * @format duration
   */
  doorlooptijd?: string | null;
  /**
   * toelichting
   * Een eventuele toelichting op dit STATUSTYPE.
   * @maxLength 1000
   */
  toelichting?: string | null;
  /** checklistitem */
  checklistitemStatustype?: CheckListItem[];
  /**
   * eigenschappen
   * de EIGENSCHAPpen die verplicht een waarde moeten hebben gekregen, voordat een STATUS van dit STATUSTYPE kan worden gezet.
   * @uniqueItems true
   */
  eigenschappen?: (string | null)[];
  /**
   * datum begin geldigheid
   * De datum waarop het is ontstaan.
   * @format date
   */
  beginGeldigheid?: string | null;
  /**
   * datum einde geldigheid
   * De datum waarop het is opgeheven.
   * @format date
   */
  eindeGeldigheid?: string | null;
  /**
   * datum begin object
   * De datum waarop de eerst versie van het object ontstaan is.
   * @format date
   */
  beginObject?: string | null;
  /**
   * datum einde object
   * De datum van de aller laatste versie van het object.
   * @format date
   */
  eindeObject?: string | null;
}
// ...
export interface StatusType {
  /**
   * url
   * URL-referentie naar dit object. Dit is de unieke identificatie en locatie van dit object.
   * @format uri
   * @minLength 1
   * @maxLength 1000
   */
  url: string;
  /**
   * omschrijving
   * Een korte, voor de initiator van de zaak relevante, omschrijving van de aard van de STATUS van zaken van een ZAAKTYPE.
   * @maxLength 80
   */
  omschrijving: string;
  /**
   * omschrijving generiek
   * Algemeen gehanteerde omschrijving van de aard van STATUSsen van het STATUSTYPE
   * @maxLength 80
   */
  omschrijvingGeneriek?: string;
  /**
   * statustekst
   * De tekst die wordt gebruikt om de Initiator te informeren over het bereiken van een STATUS van dit STATUSTYPE bij het desbetreffende ZAAKTYPE.
   * @maxLength 1000
   */
  statustekst?: string;
  /**
   * Zaaktype
   * URL-referentie naar het ZAAKTYPE van ZAAKen waarin STATUSsen van dit STATUSTYPE bereikt kunnen worden.
   * @format uri
   */
  zaaktype: string;
  /**
   * catalogus
   * @format uri
   */
  catalogus: string;
  /**
   * Zaaktype
   * Unieke identificatie van het ZAAKTYPE binnen de CATALOGUS waarin het ZAAKTYPE voorkomt.
   */
  zaaktypeIdentificatie: string;
  /**
   * statustypevolgnummer
   * Een volgnummer voor statussen van het STATUSTYPE binnen een zaak.
   * @min 1
   * @max 9999
   */
  volgnummer: number;
  /**
   * isEindstatus
   * Geeft aan dat dit STATUSTYPE een eindstatus betreft. Dit gegeven is afgeleid uit alle STATUSTYPEn van dit ZAAKTYPE met het hoogste volgnummer.
   */
  isEindstatus: boolean;
  /**
   * informeren
   * Aanduiding die aangeeft of na het zetten van een STATUS van dit STATUSTYPE de Initiator moet worden geïnformeerd over de statusovergang.
   */
  informeren?: boolean;
  /**
   * doorlooptijd status
   * De door de zaakbehandelende organisatie(s) gestelde norm voor de doorlooptijd voor het bereiken van STATUSsen van dit STATUSTYPE bij het desbetreffende ZAAKTYPE.
   * @format duration
   */
  doorlooptijd?: string | null;
  /**
   * toelichting
   * Een eventuele toelichting op dit STATUSTYPE.
   * @maxLength 1000
   */
  toelichting?: string | null;
  /** checklistitem */
  checklistitemStatustype?: CheckListItem[];
  /**
   * eigenschappen
   * de EIGENSCHAPpen die verplicht een waarde moeten hebben gekregen, voordat een STATUS van dit STATUSTYPE kan worden gezet.
   * @uniqueItems true
   */
  eigenschappen?: (string | null)[];
  /**
   * datum begin geldigheid
   * De datum waarop het is ontstaan.
   * @format date
   */
  beginGeldigheid?: string | null;
  /**
   * datum einde geldigheid
   * De datum waarop het is opgeheven.
   * @format date
   */
  eindeGeldigheid?: string | null;
  /**
   * datum begin object
   * De datum waarop de eerst versie van het object ontstaan is.
   * @format date
   */
  beginObject?: string | null;
  /**
   * datum einde object
   * De datum van de aller laatste versie van het object.
   * @format date
   */
  eindeObject?: string | null;
}

/** Formaat van HTTP 4xx en 5xx fouten. */
export interface ValidatieFout {
  /**
   * type
   * URI referentie naar het type fout, bedoeld voor developers
   */
  type?: string;
  /**
   * code
   * Systeemcode die het type fout aangeeft
   */
  code: string;
  /**
   * title
   * Generieke titel voor het type fout
   */
  title: string;
  /**
   * status
   * De HTTP status code
   */
  status: number;
  /**
   * detail
   * Extra informatie bij de fout, indien beschikbaar
   */
  detail: string;
  /**
   * instance
   * URI met referentie naar dit specifiek voorkomen van de fout. Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.
   */
  instance: string;
  /** invalidParams */
  invalidParams: FieldValidationError[];
}
// ...
export interface StatustypeListParams {
  /**
   * URL-referentie naar het ZAAKTYPE van ZAAKen waarin STATUSsen van dit STATUSTYPE bereikt kunnen worden.
   * @format uri
   */
  zaaktype?: string;
  /** zaaktype_identificatie */
  zaaktypeIdentificatie?: string;
  /**
   * filter objects depending on their concept status:
   * * `alles`: Toon objecten waarvan het attribuut `concept` true of false is.
   * * `concept`: Toon objecten waarvan het attribuut `concept` true is.
   * * `definitief`: Toon objecten waarvan het attribuut `concept` false is (standaard).
   */
  status?: string;
  /** filter objecten op hun geldigheids datum. */
  datumGeldigheid?: string;
  /** Een pagina binnen de gepagineerde set resultaten. */
  page?: number;
}

export type StatustypeListData = PaginatedStatusTypeList;

export type StatustypeCreateData = StatusType;

export type StatustypeRetrieveData = StatusType;

export type StatustypeUpdateData = StatusType;

export type StatustypePartialUpdateData = StatusType;

export type StatustypeDestroyData = any;

export type StatustypeHeadersData = any;